Add locale.

+++ zones.php
@@ -8,45 +8,35 @@ if (isset($_SESSION['zone'])) {
 }
 echo '<div class="popcontainer" id="zonepopcontainer">';
 echo '<div class="popsubcontainer">';
-echo '<div class="areas">';
+
 $zones = mysqli_query($dataconection, "SELECT * FROM area WHERE company_ID_company =" . $_SESSION['company']);
+$i = 0;
 while ($zone = mysqli_fetch_array($zones)) {
-    echo '<div class="table_row"><a class="link area" href="?z=' . $zone['ID_area'] . '">' . $zone['Area_name'] . '</a> <a class="link area" href="?z=' . $zone['ID_area'] . '&action=1">vybrat</a> <!--<a class="link area" href="?z=' . $zone['ID_area'] . '&action=2">--> upravit <!--</a>--><!--<a class="link area" href="?z=' . $zone['ID_area'] . '&action=2">--> smazat <!--</a>--></div>';
+    if ($i == 0 OR $i % $COUNTONPAGE == 0) {
+        echo '<div class="rTable areas" id="arealist'.($i/$COUNTONPAGE+1).'" style="';
+        if ($i != 0){echo 'display:none';};
+        echo '">';
+    }
+    echo '<div class="rTableRow row"><div class="rTableCell areacell" onclick="location=\'' . $_SERVER['SERVER_ROOT'] . '?z=' . $zone['ID_area'] . '\'">' . $zone['Area_name'] . '</div> <div class="rTableCell areacell" onclick="location=\'' . $_SERVER['SERVER_ROOT'] . '?z=' . $zone['ID_area'] . '\'">prejmenovat</div><div class="rTableCell areacell" onclick="if(confirm(\'' . _("Opravdu chcete smazat zónu?") . '\')){console.log(\'' . $zone['ID_area'] . '\')}else{console.log(\'Ne\')}">smazat </div></div>';
+    $i++;
+    if ($i % $COUNTONPAGE == 0) {
+        echo '</div>';
+    }
+
 }
-echo '</div>';
-echo '<div onclick="document.getElementById(\'zonepopcontainer\').style.display = \'none\';" style="
+
+for($stranka=1;$stranka <= round($i/$COUNTONPAGE);$stranka++) {
+    echo '<span ';
+    if($stranka==1) {
+        echo 'class="arealistpageactive" page='.$stranka.' ';
+    }
+    echo 'id="arealistpage'.$stranka.'" onclick="document.getElementById(\'arealist\'+document.getElementsByClassName(\'arealistpageactive\')[0].id.slice(12,document.getElementsByClassName(\'arealistpageactive\')[0].id.length)).style.display = \'none\';document.getElementsByClassName(\'arealistpageactive\')[0].className = \'\';document.getElementById(\'arealistpage'.$stranka.'\').className=\'arealistpageactive\';document.getElementById(\'arealist'.$stranka.'\').style.display =\'block\';">'.$stranka.'</span>';
+}
+echo '<div onclick="document.getElementById(\'zonepopcontainer\').style.display = \'none\';document.getElementById(\'map\').style.display = \'block\';" style="
     cursor: pointer;
 ">Zavřít</div>';
 echo '</div>';
 echo '</div>';
-//test end
+echo '</div>';
 
-//old tree areas
-//while ($parentzoneid) {
-//    $parentzone = mysqli_query($dataconection, "SELECT * FROM area WHERE ID_area =" . $parentzoneid);
-//
-//    $parent = mysqli_fetch_array($parentzone);
-//    $parents[$zonelevel] = '<a class="link area" href="?z=' . $parent['ID_area'] . '">' . $parent['Area_name'] . '</a>';
-//    if ($parentzoneid != $DEFAULTAREAID) {
-//        $parentzoneid = $parent['area_ID_area'];
-//    } else {
-//        $parentzoneid = False;
-//    }
-//    $zonelevel = $zonelevel + 1;
-//}
-//$zonelevel = $zonelevel - 1;
-//while ($zonelevel > 0) {
-//    echo $parents[$zonelevel];
-//    echo " > ";
-//    $zonelevel = $zonelevel - 1;
-//}
-//if (!isset($_SESSION['zoneshow'])) {
-//    echo "<ul>";
-//    $zones = mysqli_query($dataconection, "SELECT * FROM area WHERE area_ID_area =" . $zoneid);
-//    while ($zone = mysqli_fetch_array($zones)) {
-//        echo '<li><a class="link area" href="?z=' . $zone['ID_area'] . '">' . $zone['Area_name'] . '</a> <a class="link area" href="?z=' . $zone['ID_area'] . '&action=1">vstoupit</a> <!--<a class="link area" href="?z=' . $zone['ID_area'] . '&action=2">--> editovat(neaktivni) <!--</a>--><!--<a class="link area" href="?z=' . $zone['ID_area'] . '&action=2">--> odstranit(neaktivni) <!--</a>--></li>';
-//    }
-//    echo '<!-- <a class="link area" href="?z=new&action=3"> -->pridat(neaktivni)<!--</a>--></ul><br></div>';
-//}
-//end old tree areas
 ?>
